+++ tests/testthat/test-utils.R
@@ -109,13 +109,13 @@ test_that("remove_class_cond works", {
 # compute_def_names -------------------------------------------------------
 test_that("compute_def_names works", {
   expect_identical(compute_def_names(0), character(0))
-  expect_identical(compute_def_names(10), paste0("..", seq_len(10)))
+  expect_identical(compute_def_names(10), paste0("__", seq_len(10)))
   expect_identical(compute_def_names(10, "base"),
-                   paste0("base..", seq_len(10)))
+                   paste0("base__", seq_len(10)))
   expect_identical(compute_def_names(10, .start_ind = 4),
-                   paste0("..", seq_len(10) + 3))
+                   paste0("__", seq_len(10) + 3))
   expect_identical(compute_def_names(10, "base", 4),
-                   paste0("base..", seq_len(10) + 3))
+                   paste0("base__", seq_len(10) + 3))
 })
 
 
@@ -124,7 +124,7 @@ test_that("enhance_names works", {
   expect_identical(enhance_names(character(0)), character(0))
 
   input <- c("", "name", "", "name", "var")
-  output_ref_1 <- c("..1", "name", "..3", "name", "var")
+  output_ref_1 <- c("__1", "name", "__3", "name", "var")
 
   expect_identical(enhance_names(input), output_ref_1)
   expect_identical(
@@ -142,10 +142,10 @@ test_that("enhance_names works", {
 
   expect_identical(
     enhance_names(input, .root = "base"),
-    c("base..1", "name", "base..3", "name", "var")
+    c("base__1", "name", "base__3", "name", "var")
   )
   expect_identical(
     enhance_names(input, .root = "base", .start_ind = 5),
-    c("base..5", "name", "base..7", "name", "var")
+    c("base__5", "name", "base__7", "name", "var")
   )
 })