+++ NEWS.md
@@ -1,6 +1,12 @@
 # ruler 0.1.4.9000
 
-* Reaction to `dplyr` 0.8.0:
+This version is reaction to changes in `dplyr` 0.8.0.
+
+* Breaking changes:
+    * Name repair in `rules()` now uses `__` instead of `..` as separator for
+    function position in input `...`. This is done because of new `dplyr` name
+    repair rules which assume that `..{[0-9]}` in the end of the name can be
+    removed.
     * `rules()` behaviour now depends on version of `dplyr`. For version less
     than 0.8.0 it is a direct wrapper for `dplyr::funs()` which does custom
     name repair. For newer versions it quotes elements in `...` (except explicit